% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{acs_poverty}
\alias{acs_poverty}
\title{US County Poverty Rates by Age Group}
\format{
A tibble with 9,666 rows and 4 columns.
}
\source{
American Community Survey 2023 ACS 5-year estimates, Table B17018.
}
\usage{
acs_poverty
}
\description{
A dataset of US poverty rates by selected age groups within counties.
}
\details{
\itemize{
\item fips. County FIPS code.
\item age_group. Adults 18-64, Children <18, Seniors 65+.
\item age_rate. Poverty rate (percent) for `age_group` in county.
\item total_rate. Poverty rate (percent) for all ages in county.
}
}
\keyword{datasets}
