% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{acs_poverty_lon}
\alias{acs_poverty_lon}
\title{US County Poverty Rates by Age Group, Longer Version}
\format{
A tibble with 12,888 rows and 3 columns.
}
\source{
American Community Survey 2023 ACS 5-year estimates, Table B17018.
}
\usage{
acs_poverty_lon
}
\description{
A dataset of US poverty rates by selected age groups within counties.
}
\details{
\itemize{
\item geoid. County FIPS code.
\item age_group. Adults 18-64, Children <18, Seniors 65+, All Ages.
\item prop_poor. Proportion (0-1) of `age_group` in poverty within county.
}
}
\keyword{datasets}
