% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{asasec}
\alias{asasec}
\title{American Sociological Association Section Membership}
\format{
## `asasec`
A data frame with 52 rows and 9 columns:
\describe{
  \item{Section}{Section name.}
  \item{Sname}{Short name.}
  \item{Beginning}{Cash on hand at beginning of year (2015).}
  \item{Revenues}{Membership revenues.}
  \item{Expenses}{Section expenses.}
  \item{Ending}{Cash on hand at end of year (2015).}
  \item{Journal}{Does the Section run a journal?}
  \item{Year}{Membership year.}
  \item{Members}{Number of members (2014).}
}
}
\source{
ASA Annual Report 2016.
}
\usage{
asasec
}
\description{
Membership and some financial information for sections
of the American Sociological Association in 2014/15.
}
\details{
Data from the American Sociological Association.
}
\author{
Kieran Healy
}
\keyword{datasets}
