% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bad_date}
\alias{bad_date}
\title{Dates in character form}
\format{
A tibble with 10 rows and two columns.
}
\source{
Chris Delcher.
}
\usage{
bad_date
}
\description{
A table of dates and observations with the date column stored as a
character string.
}
\keyword{datasets}
