% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{boomer}
\alias{boomer}
\title{Monthly Births in the US, and England & Wales}
\format{
A tibble with 1,644 rows and 6 columns.
}
\source{
UK Office of National Statistics; US Census Bureau.
}
\usage{
boomer
}
\description{
Births by month, 1933-2015 (United States) and 1938-1991 (England
& Wales)
}
\details{
The variables are as follows:

\itemize{
\item  date. Year and month. (Day is arbitrarily set to 01 for all observations,
data are monthly.)
\item  month. Month of the year (1-12).
\item n_days. The number of days in a given month/year date.
\item births. Total live births for that month.
\item  total_pop. National population estimate for that month.
\item country. United States or England & Wales.
}
}
\keyword{datasets}
