% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{center_df}
\alias{center_df}
\title{center_df}
\usage{
center_df(data, sc = FALSE, cen = TRUE)
}
\arguments{
\item{data}{A data frame or tibble}

\item{sc}{Scale the variables (default FALSE)}

\item{cen}{Center the variables on their means (default TRUE)}
}
\value{
An object of the same class as `data`, with the numeric
    columns scaled or centered as requested
}
\description{
Scale and/or center the numeric columns of a data frame or tibble
}
\details{
Takes a data frame or tibble as input and scales and/or centers
    the numeric columns. By default, centers but doesn't scale
}
\examples{
head(center_df(organdata))
}
\author{
Kieran Healy
}
