% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{color_comp}
\alias{color_comp}
\title{Plot a table of color hex values as a table of colors}
\usage{
color_comp(df)
}
\arguments{
\item{df}{data frame of color hex values}
}
\value{
Plot of table of colors
}
\description{
Plot a table of color hex values as a table of colors
}
\details{
Given a data frame of color values, plot them as swatches
}
\examples{
\donttest{
color_table
color_comp(color_table)
}
}
\author{
Kieran Healy
}
