% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{color_table}
\alias{color_table}
\title{A table of hex color values related to types of color blindness}
\format{A tibble with five rows and four columns.}
\source{
Kieran Healy
}
\usage{
color_table
}
\description{
Hex values for five default ggplot colors, with corresponding
approximations for three kinds of color blindness. Produced by the
`dichromat` package.
}
\keyword{datasets}
