% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{counties_sf}
\alias{counties_sf}
\title{US County geographies}
\format{
## `counties_sf`
A data frame with 3,144 rows and 3 columns:
\describe{
  \item{fips}{County FIPS}
  \item{state}{State name}
  \item{geometry}{Geometry}
}
}
\source{
US Census Bureau.
}
\usage{
counties_sf
}
\description{
US County map data layer
}
\details{
A spatial data frame for use with the `sf` package. Alaska and Hawaii have had their geometries scaled and shifted to the bottom left of the map area.  Alaska's Aleutian islands are not included.
}
\author{
Kieran Healy
}
\keyword{datasets}
