% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{county_comp}
\alias{county_comp}
\title{US County Components of Population Change}
\format{
A tibble with 3,144 rows and 8 columns.
}
\source{
US Census Bureau Components of Population Change 2023 estimates.
}
\usage{
county_comp
}
\description{
A dataset of components of population change (rates only) for US Counties in 2023.
}
\details{
\itemize{
\item fips. County FIPS code.
\item county. County name.
\item state. State abbreviation.
\item rbirth. Birth rate.
\item rdeath. Death rate.
\item rnatchg. Natural change rate.
\item rintl. International migration rate.
\item rdom. Domestic migration rate.
\item rnetmig. Net migration rate.
}
}
\keyword{datasets}
