% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{county_data}
\alias{county_data}
\title{Census Data on US Counties}
\format{
A data frame with 3195 rows and 13 columns.
}
\source{
US Census Bureau, Centers for Disease Control
}
\usage{
county_data
}
\description{
Selected county data (including US and state-level observations on some
variables). Preserved for use with the first edition of the book only.
}
\details{
The variables are as follows:
\itemize{
\item id. FIPS State and County code (character)
\item name. State or County Name
\item state. State abbreviation
\item census_region. Census region
\item pop_dens. Population density per square mile, 2014 estimate
(seven categories).
\item pct_black. Percent black population, 2014 estimate (seven
category factor)
\item su_gun6. Firearm-related suicides per 100,000 population,
1999-2015. Factor variable cut into six categories. Note that the
values in this variable contain an inaccurate bottom-quartile
coding by construction. Do not present this variable as an
accurate measure of the firearm-related suicide rate.
}
}
\keyword{datasets}
