% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{county_map}
\alias{county_map}
\title{US County map file}
\format{"A data frame with 191,372 rows and 7 columns."}
\source{
Eric Celeste
}
\usage{
county_map
}
\description{
US county map data
}
\details{
\itemize{
\item long. Longitude
\item lat. Latitude
\item order. Order
\item hole. Hole (true/false)
\item piece. Piece
\item group. Group
\item id. FIPS code
}
}
\keyword{datasets}
