% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{edu}
\alias{edu}
\title{Years of school completed by people 25 years and over in the US.}
\format{
A tibble with 366 rows and 11 columns.
}
\source{
US Census Bureau
}
\usage{
edu
}
\description{
Counts of educational attainment (in thousands) from 1940 to 2016
}
\details{
The variables are as follows:
\itemize{
\item age Character. Cut into 25-34, 35-54, 55>
\item sex Character. Male, Female.
\item year Integer.
\item total Integer. Total in thousands.
\item elem4 Double. 0 to 4 years of Elementary School completed.
\item elem8 Double. 5 to 8 years of Elementary School completed.
\item hs3 Double. 1 to 3 years of High School completed.
\item hs4 Double. 4 years of High School completed.
\item coll3 Double. 1 to 3 years of College completed.
\item coll4 Double. 4 or more years of College completed.
\item median Double. Median years of education.
}
}
\keyword{datasets}
