% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{election24}
\alias{election24}
\title{US Presidential Election 2024, State-level results}
\format{
## `election24`
A data frame with 51 rows and 20 columns:
\describe{
  \item{state}{State name.}
  \item{st}{State abbreviation.}
  \item{fips}{State FIPS code (character).}
  \item{total_vote}{Total votes case.}
  \item{vote_margin}{Vote margin (Trump positive values; Harris negative.)}
  \item{winner}{Winning candidate.}
  \item{party}{Winning party.}
  \item{pct_margin}{Winner's percentage margin (proportion of total vote)}
  \item{r_points}{Percentage point difference between Trump vote percent and Harris vote percent}
  \item{d_points}{Percentage point difference between Harris vote percent and Trump vote percent}
  \item{pct_harris}{Harris vote share (proportion)}
  \item{pct_trump}{Trump vote share (proportion)}
  \item{pct_other}{Other vote share (proportion)}
  \item{harris_vote}{Harris vote total}
  \item{trump_vote}{Trump vote total}
  \item{other_vote}{Other vote total}
  \item{ev_dem}{Electoral votes for Harris}
  \item{ev_rep}{Electoral votes for Trump}
  \item{ev_other}{Electoral votes for Others}
  \item{census}{Census region}
}
}
\source{
Vote data from Wikipedia, https://en.wikipedia.org/wiki/2024_United_States_presidential_election
}
\usage{
election24
}
\description{
State-level vote totals and shares for the 2024 US Presidential election.
}
\author{
Kieran Healy
}
\keyword{datasets}
