% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{election24_county_df}
\alias{election24_county_df}
\title{US County-level Presidential Election data, 2024}
\format{
## `election24_county_df`
A tibble object with 3,153 rows and 6 columns:
\describe{
  \item{county_fips}{County FIPS code.}
  \item{votes_dem}{Votes for Harris/Walz ticket.}
  \item{votes_gop}{Votes for Trump/Vance ticket.}
  \item{total_votes}{Total votes cast.}
  \item{winner}{Winning party.}
  \item{flipped}{Did the party winner change from the winner in 2020? (Yes/No)}
}
}
\source{
Election data derived from https://doi.org/10.7910/DVN/VOQCHQ
}
\usage{
election24_county_df
}
\description{
A tibble with US presidential election data
}
\details{
A tibble.
}
\author{
Kieran Healy
}
\keyword{datasets}
