% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{elections24_county_sf}
\alias{elections24_county_sf}
\title{US County-level Presidential Election data, 2024}
\format{
## `elections24_county_sf`
A simple features object with 3,154 rows and 6 columns:
\describe{
  \item{county_fips}{County FIPS code.}
  \item{total_votes}{Total votes cast.}
  \item{votes_dem}{Votes for Harris/Walz ticket.}
  \item{votes_gop}{Votes for Trump/Vance ticket.}
  \item{winner}{Winning party.}
  \item{geometry}{Geometry.}
}
}
\source{
Geometries from US Census Bureau; vote data from Tony McGovern, `https://github.com/tonmcg/US_County_Level_Election_Results_08-24`
}
\usage{
elections24_county_sf
}
\description{
A simple features object with US presidential election data
}
\details{
A simple features object. Load the `sf` package before using. Alaska and Hawaii have had their geometries scaled and shifted to the bottom left of the map area. Alaska's Aleutian islands are not included. Alaska boundaries are for Alaskan House Districts, not counties.
}
\author{
Kieran Healy
}
\keyword{datasets}
