% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fredts}
\alias{fredts}
\title{Monetary Base and S&P 500 series}
\format{
A data frame with 5 columns and 357 rows.
}
\source{
FRED data.
}
\usage{
fredts
}
\description{
Two time series of financial data from FRED, the _i means indexed
to 100 in the base observation.
}
\keyword{datasets}
