% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{freq_tab}
\alias{freq_tab}
\title{Generate a tidy n-way frequency table}
\usage{
freq_tab(df, ...)
}
\arguments{
\item{df}{tibble or data frame (implicit within pipline)}

\item{...}{grouping, as with group_by()}
}
\value{
A tibble with the grouping variables, the N (`n`) per group, and
    the proportion (`prop`) of each group, calculated with respect to the
    outermost grouping variable.
}
\description{
Generate a tidy n-way frequency table
}
\details{
Tidyverse, pipeline, and dplyr-friendly frequency tables
}
\examples{
mtcars \%>\% freq_tab(vs, gear, carb)

}
\author{
Kieran Healy
}
