% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gss_sib}
\alias{gss_sib}
\title{General Social Survey data, 1972-2016}
\format{A data frame with 60,423 rows and 7 variables.}
\source{
National Opinion Research Center, http://gss.norc.org.
}
\usage{
gss_sib
}
\description{
A dataset containing an extract from the General Social
Survey. See http://gss.norc.org/Get-Documentation for full
documentation of the variables. This data contains seven
variables from `gss_lon` with all NA values omitted.
}
\details{
\itemize{
\item year. gss year for this respondent.
\item id. respondent id number.
\item age. age of respondent.
\item race. race of respondent.
\item sex. respondent's sex.
\item siblings. Number of brothers and sisters (recoded from SIBS).
\item kids. Number of children (recoded from CHILDS).
}
}
\keyword{datasets}
