% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{is_html_output}
\alias{is_html_output}
\title{is_html_output}
\usage{
is_html_output(fmt = pandoc_to(), excludes = NULL)
}
\arguments{
\item{fmt}{format}

\item{excludes}{excludes}
}
\value{
html flag
}
\description{
Check if something is html output
}
\details{
internal knitr function
}
\author{
Yihui Xie
}
