% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{margin_codenote}
\alias{margin_codenote}
\title{margin_codenote}
\usage{
margin_codenote(text, icon = "&#8853;")
}
\arguments{
\item{text}{The text of the note}

\item{icon}{An icon}
}
\value{
A marginal note in monospace font
}
\description{
Use in inline R expressions to write a monospaced margin note for tufte-latex documents
}
\details{
Adapted from Dirk Eddelbuettel's tint package.
}
\author{
Kieran Healy
}
