% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{margin_textnote}
\alias{margin_textnote}
\title{margin_textnote}
\usage{
margin_textnote(text, icon = "&#8853;")
}
\arguments{
\item{text}{The text of the note}

\item{icon}{An icon}
}
\value{
A marginal note
}
\description{
Use in inline R expressions to write a margin note for tufte-latex documents
}
\details{
Borrowed from Dirk Eddelbuettel's tint package.
}
\author{
Kieran Healy (but really Dirk Eddelbuettel)
}
