% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{marginnote_html}
\alias{marginnote_html}
\title{marginnote_html}
\usage{
marginnote_html(text = "", icon = "&#8853;")
}
\arguments{
\item{text}{The text of the note}

\item{icon}{An icon}
}
\value{
An HTML note
}
\description{
Marginal note for tufte html output
}
\details{
Make a html margin note
}
\examples{
\dontrun{
marginnote_html("Hello")
}
}
\author{
Dirk Eddelbuettel
}
