% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{maunaloa}
\alias{maunaloa}
\title{Mauna Loa Atmospheric CO2 Concentration}
\format{
A data frame with 4 columns and 271 rows.
}
\source{
R base datasets; Cleveland (1993).
}
\usage{
maunaloa
}
\description{
A subset of the co2 data in base R's [datasets] package, in
a ggplot2-friendly format.
}
\keyword{datasets}
