% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{multiplot}
\alias{multiplot}
\title{multiplot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{ggplot objects}

\item{file}{.}

\item{cols}{Number of columns in the layout. Ignored if layout is present.}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\value{
Prints a grid of plot objects
}
\description{
Arrange ggplot objects on an arbitrary page grid
}
\details{
ggplot objects can be passed in ..., or to plotlist (as a list of
    ggplot objects). If the layout is something like matrix(c(1,2,3,3), nrow=2,
    byrow=TRUE), then plot 1 will go in the upper left, 2 will go
    in the upper right, and 3 will go all the way across the bottom.
}
\author{
Winston Chang
}
