% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{\%nin\%}
\alias{\%nin\%}
\title{\%nin\%}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{vector of items}

\item{y}{vector of all values}
}
\value{
logical vecotor of items in x not in y
}
\description{
Convenience not-in operator
Complement of the built-in operator \code{\%in\%}. Returns the elements of \code{x} that are not in \code{y}.
}
\examples{
fruit <- c("apples", "oranges", "banana")
"apples" \%nin\% fruit
"pears" \%nin\% fruit
}
\author{
Kieran Healy
}
