% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nyc_zips}
\alias{nyc_zips}
\title{New York City Zip Code tabulation areas}
\format{A Simple feature collection with 262 features and 11 fields}
\source{
New York City
}
\usage{
nyc_zips
}
\description{
A simple features (sf) object with zip code boundary information
for New York City.
}
\details{
The variables are as follows:
\itemize{
\item objectid (integer) Record ID
\item zip_code (integer) Zip code
\item po_name (integer) Post Office name
\item state (charater) State
\item borough (character) City borough
\item st_fips (character) State FIPS code
\item cty_fips (character) County FIPS code
\item bld_gpostal code Building postal code
\item shape_leng (double)
\item shape_area (double)
\item id (character)
\item geometry (polygon)
}
}
\keyword{datasets}
