% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{oecd_le}
\alias{oecd_le}
\title{Life Expectancy in the OECD, 1960-2015.}
\format{"A tibble with 1,746 rows and 4 columns."}
\source{
OECD
}
\usage{
oecd_le
}
\description{
Life expectancy data for individual countries.
}
\details{
The variables are as follows:

\itemize{
\item  country. Country. (Character)
\item  year. Year. (Integer.)
\item  lifeexp. Life Expectancy at Birth, measured in years.
\item  is_usa. Indicator for USA or Other country.
}
}
\keyword{datasets}
