% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{oecd_sum}
\alias{oecd_sum}
\title{Life Expectancy in the OECD, 1960-2015}
\format{"A tibble with 57 rows and 5 columns."}
\source{
OECD
}
\usage{
oecd_sum
}
\description{
Life expectancy data summary table.
}
\details{
The variables are as follows:

\itemize{
\item  year. Year. (Integer.)
\item  other. Life Expectancy at birth in OECD countries excluding
the USA. Measured in years.
\item usa. Life Expectancy at birth in the USA. Measured in years.
\item  diff. Difference between usa and other.
\item hi_lo. Is usa above or below the oecd average?
}
}
\keyword{datasets}
