% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{okboomer}
\alias{okboomer}
\title{Monthly Births in the U.S., 1933-2015}
\format{
## `okboomer`
A data frame with 996 rows and 11 columns:
\describe{
  \item{date}{Date in date format}
  \item{year_fct}{Year as ordered factor}
  \item{month_fct}{Month as ordered factor}
  \item{n_days}{N of days in this month}
  \item{births}{Total births in this month}
  \item{total_pop}{Population}
  \item{births_pct}{Births as a proportion of total population}
  \item{births_pct_day}{Average daily births per million population}
  \item{seasonal}{Seasonal component from an STL decomposition of `births_pct_day`}
  \item{trend}{Trend component from an STL decomposition of `births_pct_day`}
  \item{remainder}{Remainder component from an STL decomposition of `births_pct_day`}
}
}
\source{
U.S. Census Bureau.
}
\usage{
okboomer
}
\description{
Births by month, 1933-2015, with decomposition components.
}
\details{
Dataset originally constructed to reproduce a visualization exercise by Aaron Penne.
}
\author{
Kieran Healy
}
\keyword{datasets}
