% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{opiates}
\alias{opiates}
\title{Opiate-Related Deaths in the United States, 1999-2020}
\format{
## `opiates`
A tibble frame with 1,122 rows and 8 columns:
\describe{
  \item{fips}{State FIPS code.}
  \item{st}{State abbreviation.}
  \item{year}{Year.}
  \item{deaths}{N opiate-related deaths.}
  \item{crude}{Crude death rate per 100,000 population.}
  \item{adjusted}{Adjusted death rate.}
  \item{region}{Census region.}
  \item{division_name}{Census division.}
}
}
\source{
CDC WONDER, http://wonder.cdc.gov/mcd-icd10.html
}
\usage{
opiates
}
\description{
State-level data on opiate related deaths in the US, from the CDC WONDER database.
}
\details{
Dataset is Multiple Cause of Death, 1999-2020. Standard Population: 2000 U.S. Std. Population. Rates per 100,000. Default intercensal populations for years 2001-2009. MCD ICD-10 Codes selected: T40.0 (Opium), T40.1 (Heroin), T40.2 (Other opioids), T40.3 (Methadone), T40.4 (Other synthetic narcotics), T40.6 (Other and unspecified narcotics). UCD ICD-10 Codes selected: X40-X44, X60-X64, X85, Y10-Y14.
}
\author{
Kieran Healy
}
\keyword{datasets}
