% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{out_format}
\alias{out_format}
\title{out_format}
\usage{
out_format(x)
}
\arguments{
\item{x}{x}
}
\value{
format
}
\description{
out_format
}
\details{
internal knitr function
}
\author{
Yihui Xie
}
