% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{pandoc_to}
\alias{pandoc_to}
\title{pandoc_to}
\usage{
pandoc_to(x)
}
\arguments{
\item{x}{fmt}
}
\value{
opt_knit element
}
\description{
pandoc_to
}
\details{
what are we converting to
}
\author{
Yihui Xie
}
