% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{preg2}
\alias{preg2}
\title{A wider table of untidy data}
\format{
An tbl_df object (inherits from tbl, data.frame) with 2 rows and 4 columns.
}
\source{
Hadley Wickham (2014).
}
\usage{
preg2
}
\description{
A second table of data from Wickham (2014).
}
\keyword{datasets}
