% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{round_df}
\alias{round_df}
\title{round_df}
\usage{
round_df(data, dig = 2)
}
\arguments{
\item{data}{A data frame or tibble}

\item{dig}{The number of digits to round to}
}
\value{
An object of the same class as `data`, with the numeric
    columns rounded off to `dig`
}
\description{
Round numeric columns of a data frame or tibble
}
\details{
Takes a data frame or tibble as input, rounds the numeric columns to the
    specified number of digits.
}
\author{
Kieran Healy
}
