% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{setup_course_notes}
\alias{setup_course_notes}
\title{setup_course_notes}
\usage{
setup_course_notes(folder = "~/Desktop",
  zipfile = "dataviz_course_notes.zip",
  packet = "dataviz_course_notes")
}
\arguments{
\item{folder}{The destination to copy to within the user's home.
By default ~/Desktop.}

\item{zipfile}{The name of the zipped course materials file in the
socviz library.}

\item{packet}{The name of the course packet folder to be created}
}
\value{
The `zipfile` is copied to `folder` and its contents
    expanded into a directory, the `packet`.
}
\description{
Copy and expend course notes to the desktop
}
\details{
Transfers a zip file containing course materials from the socviz
    library to the Desktop.
}
\author{
Kieran Healy
}
