% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{states_sf}
\alias{states_sf}
\title{US State geometries}
\format{
## `states_sf`
A simple features object with 51 rows and 5 columns:
\describe{
  \item{fips}{State FIPS code}
  \item{st}{State name abbreviation}
  \item{state}{State name}
  \item{census}{Census region}
  \item{geometry}{Geometry}
}
}
\source{
US Census Bureau.
}
\usage{
states_sf
}
\description{
US State map data layer
}
\details{
A simple features object. Load the `sf` package before using. Alaska and Hawaii have had their geometries scaled and shifted to the bottom left of the map area. Alaska's Aleutian islands are not included.
}
\author{
Kieran Healy
}
\keyword{datasets}
