% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{studebt}
\alias{studebt}
\title{Student debt data}
\format{
## `studebt`
A data frame with 16 rows and 4 columns:
\describe{
  \item{Debt}{Debt categories (character)}
  \item{type}{Pct in terms of Borrowers or Balances}
  \item{pct}{Percentage of all type}
  \item{Debtrc}{Debt categories (ordered factor)}
}
}
\source{
Federal Reserve Bank of New York.
}
\usage{
studebt
}
\description{
Outstanding student debts in 2016 across 8 income categories, by
percent of all borrowers and percent of all balances.
}
\keyword{datasets}
