% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{theme_map}
\alias{theme_map}
\title{theme_map}
\usage{
theme_map(base_size = 9, base_family = "")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\value{
A ggplot theme layout
}
\description{
Mostly blank ggplot2 theme for use with maps
}
\details{
A ggplot2 theme
}
\author{
Bob Rudis
}
