% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{titanic}
\alias{titanic}
\title{A table of survival rates from the Titanic}
\format{
A data frame with four rows and four columns.
}
\source{
Titanic data
}
\usage{
titanic
}
\description{
A small table of survival rates from the Titanic, by sex
}
\keyword{datasets}
