% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{yahoo}
\alias{yahoo}
\title{Yahoo Revenue and Employees}
\format{
A tibble with 4 columns and 12 rows.
}
\source{
QZ.com
}
\usage{
yahoo
}
\description{
Data on Revenue and Employees at Yahoo before and during Marissa
Mayer's tenure as CEO.
}
\keyword{datasets}
