% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{asasec}
\alias{asasec}
\title{American Sociological Association Section Membership}
\format{A data frame with 572 rows and 9 columns.}
\source{
ASA Annual Report 2016
}
\usage{
asasec
}
\description{
Membership (2005-2015) and some financial information for sections
of the American Sociolgical Association
}
\keyword{datasets}
