% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{color_pal}
\alias{color_pal}
\title{Draw a palette of colors}
\usage{
color_pal(col, border = "light gray", ...)
}
\arguments{
\item{col}{vector of colors}

\item{border}{border}

\item{...}{other arguments}
}
\value{
Plot of a color palette
}
\description{
Draw a palette of colors
}
\details{
Borrowed from the colorspace library
}
\author{
colorspace library authors
}
