% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{counties_sf}
\alias{counties_sf}
\title{US County geometries and demographic data}
\format{
## `counties_sf`
A simple features object 3,144 rows and 15 columns:
\describe{
  \item{fips}{FIPS code.}
  \item{name}{County name.}
  \item{area_sqmi}{Area in square miles.}
  \item{white}{N White population.}
  \item{black}{N Black population.}
  \item{asian}{N Asian population.}
  \item{nh_white}{N Non-Hispanic White population.}
  \item{hispanic}{N Hispanic population.}
  \item{pop}{Total population.}
  \item{black_disc}{Percent Black, discretized.}
  \item{hisp_disc}{Percent Hispanic, discretized.}
  \item{nhwhite_disc}{Percent Non-Hispanic White, discretized.}
  \item{asian_disc}{Percent Asian, discretized.}
  \item{pop_dens}{Population density per square mile.}
  \item{pop_dens_disc}{Population density per square mile, discretized.}
\item{su_gun6}{Firearm-related suicides per 100,000 population,
1999-2015. Factor variable cut into six categories. Note that the
values in this variable contain an inaccurate bottom-quartile
coding by construction. Do not present this variable as an
accurate measure of the firearm-related suicide rate.}
  \item{geometry}{Geometry.}
}
}
\source{
US Census Bureau.
}
\usage{
counties_sf
}
\description{
US County map data layer with selected Census Bureau demographic variables.
}
\details{
A simple features object. Load the `sf` package before using. Alaska and Hawaii have had their geometries scaled and shifted to the bottom left of the map area. Alaska's Aleutian islands are not included. Population counts and other demographic information are from the 2024 5-year ACS estimates.
}
\author{
Kieran Healy
}
\keyword{datasets}
