% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{county_data}
\alias{county_data}
\title{Census Data on US Counties}
\format{"A data frame with 3195 rows and 13 columns."}
\source{
US Census Bureau
}
\usage{
county_data
}
\description{
Selected county data (including state-level observations)
\itemize{
\item id. FIPS State and County code
\item name. State or County Name
\item state. State abbreviation
\item census_region. Census region
\item pop_dens. Population density per square mile, 2014 estimate
\item pct_black. Percent black population, 2014 estimate
\item pop. Population, 2014 estimate
\item female. Female persons, percent, 2013
\item white. White alone, percent, 2013
\item black. Black alone, percent, 2013
\item travel_time. Mean travel time to work (minutes), workers age 16+, 2009-2013
\item land_area. Land area in square miles, 2010
\item hh_income. Median household income, 2009-2013
}
}
\keyword{datasets}
