% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nyc_license}
\alias{nyc_license}
\alias{nyc_zips}
\title{Active dog licenses in New York City, 2016
This data has been partially cleaned and recoded from the original source}
\format{"A tibble with 122,203 rows and 15 columns."}
\source{
New York City OpenData: https://data.cityofnewyork.us/Health/NYC-Dog-Licensing-Dataset/nu7n-tubp

"New York City"
}
\usage{
nyc_license

nyc_zips
}
\description{
The variables are as follows:

The variables are as follows:
}
\details{
\itemize{
\item row_number (integer) Record ID
\item animal_name (character) Name of dog
\item animal_gender (character) M (Male), F (Female)
\item animal_birth_month (date) Month and year of dog's birth.
Encoded as a date, but all "day" values are the first of the month.
\item breed_rc (character) Breed of dog.
\item borough (character) City borough of owner
\item zip_code (integer) Zip code of owner
\item census_tract2010 (integer) Census Tract (2010 Census) of owner
\item nta (character) Neighborhood tabulation area of owner
\item city_council_district (integer) City council district of owner
\item congressional_district (integer) Congressional district of owner
\item state_senatorial_district (integer) State senatorial district
of owner
\item license_issued_date (date) Date license issued
\item license_expiered_date (date) Date license expired
}

\itemize{
\item objectid (integer) Record ID
\item zip_code (integer) Zip code
\item po_name (integer) Post Office name
\item state (charater) State
\item borough (character) City borough
\item st_fips (character) State FIPS code
\item cty_fips (character) County FIPS code
\item bld_gpostal code Building postal code
\item shape_leng (double)
\item shape_area (double)
\item id (character)
\item geometry (polygon)
}
}
\keyword{datasets}
