% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{setup_course_notes}
\alias{setup_course_notes}
\title{Copy and expand course notes to a specified folder}
\usage{
setup_course_notes(folder, zipfile = "dataviz_course_notes.zip",
  packet = "dataviz_course_notes")
}
\arguments{
\item{folder}{The destination to copy to within the user's home.
This must be supplied by the user.}

\item{zipfile}{The name of the zipped course materials file in the
socviz library.}

\item{packet}{The name of the course packet folder to be created}
}
\value{
The `zipfile` is copied to `folder` and its contents
    expanded into a directory, the `packet`.
}
\description{
Copy and expand course notes to a specified folder
}
\details{
Transfers a zip file containing course materials from the socviz
    library to the Desktop.
}
\examples{
\dontrun{
setup_course_notes(folder = tempdir())
}
}
\author{
Kieran Healy
}
