% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{studebt}
\alias{studebt}
\title{Student debt data}
\format{"A tibble with 16 rows and 4 columns."}
\source{
Federal Reserve Bank of New York
}
\usage{
studebt
}
\description{
Outstanding student debts in 2016 across 8 income categories, by
percent of all borrowers and percent of all balances.
}
\keyword{datasets}
